requirements_test.txt
setuptools>=41.0.0
flake8>=3.7.7
flake8-print>=3.1.0
autopep8>=1.4.3
nose>=1.3.7
coverage>=4.5.3
PyHamcrest>=1.9.0
numpy==1.16.2